requirements.txt
aliyun-python-sdk-alidns==2.6.29
aliyun-python-sdk-core==2.13.33
aliyun-python-sdk-domain==3.14.5
aliyun-python-sdk-ecs==4.23.11
amqp==2.6.1
asgiref==3.3.1
billiard==3.6.4.0
cachetools==4.2.1
celery==4.4.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.1.7
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.7.0
django-filter==2.4.0
django-log-request-id==1.6.0
django-timezone-field==4.1.2
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
dnspython==1.16.0
et-xmlfile==1.0.1
eventlet==0.30.2
google-auth==1.28.1
greenlet==1.0.0
idna==2.10
jdcal==1.4.1
jmespath==0.10.0
kombu==4.6.11
kubernetes==12.0.1
oauthlib==3.1.0
openpyxl==3.0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==1.0.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
vine==1.3.0
websocket-client==0.58.0
Pint==0.17
channels==3.0.3
ansible==2.8.8
paramiko==2.7.2
pymongo==3.11.4
django-redis==4.12.1
pywinrm>=0.2.2
demjson==2.2.4
channels-redis==3.2.0